requirements.txt
deepface
tf-keras
streamlit
pillow

